Add ws_share to easily manage read access for other users to a workspace
Users often want to give other users, e.g., support staff, read access to
their workspaces.  Often these users are not in the same user group as the
owner of the workspace. So the only way with standard UNIX perms is to grant
read access for other, which may be critical for some users.
A solution therefore is to use ACLs. However the understanding and correct
usage of ACLs is not easy.

The new ws_share command allows to easily grant and remove read access for
other users as well as list users that were granted access to a workspace.
The ws_share command will perform the necessary command to modify the ACLs
transparently for the user.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>