Add ws_delete to delete all content of a workspace and releasing it
On some systems quotas are applied to workspaces. If a user releases a
workspace with ws_release it is only released and it will still account
to the user's quota - until the workspace expirer cleans it up.
So, if a user is close to quota limits he has to delete all the content
of a workspace before releasing it.

The new ws_delete command allows to easily delete the content of a
workspace and releasing it in one command.

Note: Deletion in this version of the script is performed with rsync as
a faster method compared to a normal "rm -rf".

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>